Rewrite the packetfifo code
I don't know what the author was thinking when he wrote it, but it
was executing the idea very, very poorly.  The correct idiom is to
have a second back pointer that gets updated when the data is
declared good.  Having a second front pointer is wrong and causes
issues.